byte swap

+++ flows/photometry.py
@@ -186,8 +186,7 @@ def photometry(fileid, output_folder=None, attempt_imagematch=True, keep_diff_fi
     image.error = calc_total_error(image.clean, bkg.background_rms, 1.0)
 
     # Use sep to for soure extraction
-    image.sepdata = np.asarray(image.image)
-    #image.setpdate = np.asrray(image.image
+    image.sepdata = np.asarray(image.image,dtype=np.float64)
     image.sepbkg = sep.Background(image.sepdata, mask=image.mask)
     image.sepsub = image.sepdata - image.sepbkg
     logger.debug('sub: {} bkg_rms: {} mask: {}'.format(np.shape(image.sepsub), np.shape(image.sepbkg.globalrms),
